+++ Example_Card_Calypso/src/main/java/org/eclipse/keyple/card/calypso/example/UseCase13_PerformanceMeasurement_DistributedReloading/Main_PerformanceMeasurement_DistributedReloading_Pcsc.java
@@ -117,7 +117,6 @@ public class Main_PerformanceMeasurement_DistributedReloading_Pcsc {
         calypsoCardService
             .createCardSelection()
             .acceptInvalidatedCard()
-            .filterByCardProtocol(ConfigurationUtil.ISO_CARD_PROTOCOL)
             .filterByDfName(cardAid)
             .prepareReadRecord(
                 CalypsoConstants.SFI_ENVIRONMENT_AND_HOLDER, CalypsoConstants.RECORD_NUMBER_1)