+++ CHANGELOG.md
@@ -3,7 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [Unreleased]]
+### Fixed
+- Removed filtering by protocol in performance measurement tools.
+### Upgraded
+- "Keyple Calypso Card Library" to version `2.3.1`.
 
 ## [2022-12-13]
 ### Added